完成商户模块全部内容（接下来是用户模块）

+++ src/main/java/me/superning/luntan/service/MerchantService.java
@@ -38,7 +38,7 @@ public interface MerchantService{
      * @param id 商户id
      * @return {@link Response}
      */
-    Response buildMerchanrById(Long id);
+    Response buildMerchanrById(int id);
 
     /**
      * 投放优惠卷
